+++ rossum-agent/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 ---
 
 ## [Unreleased] - YYYY-MM-DD
+-
+
+
+## [0.2.4] - 2025-12-14
 - Added image input support [#67](https://github.com/stancld/rossum-mcp/pull/67)
 - Added logging of chat metadata into Redis for auditing [#62](https://github.com/stancld/rossum-mcp/pull/62)
 - Stopped replaying CoT in the model context [#61](https://github.com/stancld/rossum-mcp/pull/61)